cpu-o3: HTM support for nacked stores in O3 for non-TSO (ARM)
Handle retry of nacked stores whe multiple stores may be in
flight concurrently with the store that was nacked. The flag
skipWritebackReplay is used for sanity checks, but it is
synonym with 'committed' as it indicates that storeWBIt
pointer (next SQ entry to be written back) has already
been moved past this entry (e.g. store sent to cache).

Change-Id: I853b843e704a789920e1e49f15f9232123268c64