mem-ruby: Preliminary HTM support for logging (eager versioning).
Stores that target blocks that are not part of the write set
are intercepted by the TransactionalSequencer, which generates
the corresponding store accesses to the log in order to write
addr+data before the targeted block gets speculatively modified.

This patch adds basic logging functionality for filling the
log in hardware, but support for aborting an eager transaction
(unrolling the log) is not yet implemented.

Change-Id: Ie235324abeb7aa5ff6b5793d78100bbd7e5f4b3b