sim: Additional hack in pseudo inst to support logging in HTM.
Pass log base address to simulator in order to setup the log.
For now, the virtual-to-physical translations of the pages where
the log resides are collected by Ruby at the beginning of the
simulation using the hack. The returned value will indicate
whether the library code should walk the log or not. When
setting up the log at the beginning of the simulation, Ruby
will intercept accesses within the address range occupied
by the log and store the virtual-to-physical translations.

This only works as long as the OS does not remap the physical
pages where the log resides in the middle of the simulation.
However, we opted out for this solution before we figure out
an easy way to translate log address without affecting too
much the CPU code.

Change-Id: I9061373b93860138884134ee88f7c27efe9a9e29