ext: HTM library support for unrolling the log in software.
The returned abort status indicates that the memory has not
been restored to its pre-transational state, and that to do
so the undo log needs to be unrolled.
Use of the stack is precluded for all code following xbegin
since the stack frame will be overwritten as part of the
log unroll. Thus, all htm_xxxx subroutines have been
converted to macro definitions to prevent function calls.
Also, the abort handler must notify the simulator when the
log unroll has been completed, so that the hw logging
logic is reset and the TL is set back to 0.