mem-ruby: Fixed HTM protocol support for nacking remote reqs.
When L0_Nack_else received at L1 as a result of a conflict, use
global flag to detect that the remote request in the L2 request
queue must be nacked when woken up immediately after the
response from is processed. This avoid having many additional
states "*L0_N" which are only used to indicate that a L0 nack
was received.

Change-Id: Ia3549730037931e5fe2c3c1aac576c5e0a7c8c1d