cpu-o3: Fixed subtle bug related to nacked stores and ST2LD fwd.
Thanks the HTM lockstep debugger, located loads that were
obtaining the wrong data from cache by "bypassing" a logically
earlier store that should have forwarded data to the load but
was considered as out of the store buffer by the CPU while not
yet performed in cache. This happened because there is a
window of time in which there is no "pending" store in the
Sequencer that prevents later accesses to the same block from
performing in cache (i.e. later loads did not coalesce with
the pending store since there was none after the failed
callback removes all state in the Sequencer tables for the store).

Change-Id: I04f8f8790d56f219c15de45cbd5949ce5c773dfa