mem-ruby,cpu-o3: Alternative lazy HTM commit implementation.
The previous implementation invariably imposed a high penalty
on performance by squashing the pipeline on every transaction
commit, since the htmStop instruction always faulted with
ReExec as a result of getting a response from cache with
NO_FAIL_RETRY code.
This implementation moves the "retry" behaviour entirely to
Ruby by recurrently scheduling an event to retry the commit
packet. From the pov of the CPU, the htmStop instruction
(load) simply takes longer to send a response back to the CPU
(whereas HTM commands usually callback the CPU immediately)

Change-Id: I1dc41d98a1a175ebe6776bfca67cd509321ed978