mem-ruby: Major redesign of HTM eager versioning for nonTSO.
To prevent the appearance of "gaps" in the log due to
stores that cannot complete in cache being concurrent
with other stores that do add to the log, the logging
functionality was redesigned so that log requests
are made when the store is known to be able to complete.
The store is added to the Wset as soon as the store
miss is served but the hitCallback is delayed until
both log addr and log data requests are completed.
This avoids having lingering log stores from aborted
transactions and other ugly stuff

Change-Id: I2c71c7272c5de9894ad85441ddbd2548b0755470