mem: Major redesign of HTM requester stalls (retry stores)
Retrying stores that failed to perform in the memory system
due to conflicts (nacked stores) in the same way that loads
are re-executed has proven to be a *BAD* design choice: For
TSO (x86), only minor behavioural changes were required in
the LSQ logic to reuse the code in LSQUnit::writebackStore
for retrying failed (nacked) stores. But, when moving to
non-TSO system, several important risks to correctness
arouse. After struggling to get the code in LSQUnit to
tolerate the "replay" of stores that already committed
(e.g. handle cases when the blocked store is not pointed
by storeWBIt, etc.), we finally came to realize that
relying on writebackStores to retry a nacked store makes
it possible for a later load to "miss" the preceding
nacked store pending to be retried, and access cache to
find the older value, thus violating sequential program
semantics. With much pain after some tens of hours trying
to get nacked stores to work in non-TSO, including
value-based debugging with lockstep mode, etc. we have
decided that it is more suitable to have a "dual" scheme
of retrying conflicts, one for loads (involving the CPU)
and one for stores (without involving the CPU). Thus,
after this commit expect another clean-up commit in the
lsq to get rid of all the modifications to writebackStores,
removal of completeNackedStore, etc.

Change-Id: I671f3186de0722752fbfc8d0bd3875f32cb107be