mem-ruby: Handling of lingering transactional acceses.
Detect lingering transactional accesses that get to Ruby
 after the abort has completed (abort flag no longer set).
These accesses need to be cancelled and suppressed (avoid
retrying in failed callback). Allowing lingering trans
accesses to proceed gave way to wrong LLSC behaviour in
ARM due to invalidations of the lock block not being sent
to the CPU because of the abort flag mistakenly set while
no longer in transaction.
NOTE: these transactional lingering requests did not appear
earlier because the abort fault reset the htmStartStop in the
LSQUnit, but that is no longer true (reset occurs when the
pipeline squash happens, some cycles after the HTM_ABORT signal
has been sent to Ruby). For now, we opt for filtering these
lingering memory accesses in Ruby, and we believe it's best
if they are still marked as transactional and tagged with
the htmUid of the aborted transactions, which makes them
easier to detect.

Change-Id: Id460ba5c7558712b50a245e2280b5ac6f1504801