ext: Changes to outlier removal
Remove simulations with the highest execution time (as we did for huawei_closeup_presentation), which may not be those simulations with the highest absolute error from the current average. This criteria makes sense only if we assume that outliers are caused only by long context switches. It tends to remove more simulations than using the absolute error as the criteria.

Change-Id: I8a8fe71ca597d7a2087679f6f92032c0769aa71b