mem-ruby: Update LRU upon transactional writeback.
Fix performance bug leading to many "l1priv" aborts as a result
of choosing write-set blocks as victims for eviction, in spite
of having received a writeback of consistent data for the block
(due to transactional store to dirty block). The WriteBack_Copy
are an indication of data being actively accessed in L1.

Change-Id: I8c3d26dc28b02c29d74290b27e9f89ba3bdae334