mem-ruby: Fixed (performance) bug in power/req-loses due to nacks
No need for special handling for power/req-stalls, must set
the "access failed in cache" flag so that nacked mem. refs do no
retire and cause "garbage" data from the nacked access to become
visible to the program (this bug was detected because page-fault
induced aborts occured as a result of using "nacked" data. Also,
set the "transaction failed in cache" flag in the packet to
abort as soon as possible

Change-Id: Iadbe5e92726c91cbb17474d07601d72132ee2d8e