mem-ruby,cpu: HTM support for handling RMW and syscalls.
Assert failures within transactional boundaries used to cause
sanity checks in TransactionalSequencer to fail due to
unexpected Locked_RMW types. This patch refactors code
around writeCallback to detect and handle RMW accesses
inside transactions, so that the cache controller is
unblocked for the RMW address if the abort occurs
between RMW_read and RMW_write.
Such RMW accesses are often the result of calling libc
function when things go soutch (assert failure, segv, etc.)
and typically precede a syscall instruction. So this
patch also handles syscalls in the timing CPU so that
the instruction is intercepted and the transaction aborted
as if an exception (fault) occurred (non-retriable)

Change-Id: I709d0fc6da2fc51a80237257000bad7e86991cd9